fix conversion float/double
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@17871 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 768ae59f2a011a30bba8a11d767d7e2d1142fb6b

+++ Rec/Tr/TrackFitEvent/src/OTMeasurement.cpp
@@ -1,4 +1,4 @@
-// $Id: OTMeasurement.cpp,v 1.17 2008-05-27 09:51:00 wouter Exp $
+// $Id: OTMeasurement.cpp,v 1.18 2008-05-29 14:03:58 smenzeme Exp $
 // Include files
 
 // OTDet
@@ -65,9 +65,9 @@ double LHCb::OTMeasurement::stereoAngle() const
 OTDet::RadiusWithError  LHCb::OTMeasurement::driftRadiusWithError( double arclen ) const 
 {
   // Get the distance to the readout
-  float distToReadout = trajectory().endRange() - arclen;
+  double distToReadout = trajectory().endRange() - arclen;
   // Drifttime
-  float drifttime = calibratedTime() - deltaTimeOfFlight() - distToReadout / module().propagationVelocity() ;
+  double drifttime = calibratedTime() - deltaTimeOfFlight() - distToReadout / module().propagationVelocity() ;
   // The rest is done by the module
   return module().driftRadiusWithError( drifttime ) ;
 }
@@ -75,9 +75,9 @@ OTDet::RadiusWithError  LHCb::OTMeasurement::driftRadiusWithError( double arclen
 OTDet::RadiusWithError  LHCb::OTMeasurement::driftRadiusWithErrorFromY( double globalY ) const 
 {
   // Get the distance to the readout
-  float deltaY = trajectory().endPoint().y() - globalY ;
+  double deltaY = trajectory().endPoint().y() - globalY ;
   // Drifttime
-  float drifttime = calibratedTime() - deltaTimeOfFlight() - deltaY / module().propagationVelocityY() ;
+  double drifttime = calibratedTime() - deltaTimeOfFlight() - deltaY / module().propagationVelocityY() ;
   // The rest is done by the module
   return module().driftRadiusWithError( drifttime ) ;
 }
